Mobile version of features, deleting unused files
