--- src/components/LandingFeatures/features/Enclave.tsx
@@ -1,74 +1,0 @@
-import styles from '../landing-features.module.scss';
-
-const Enclave = () => {
-  return (
-    <svg
-      className={styles.feature__logo}
-      width="93"
-      height="111"
-      viewBox="0 0 93 111"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M36.7193 56.6123V73.931L28.3857 80.4036V86.8763"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <circle
-        cx="28.3861"
-        cy="90.3851"
-        r="3.88608"
-        fill="#343A3F"
-        stroke="#343A3F"
-      />
-      <path
-        d="M55.9565 56.6121V73.9307L64.29 80.4034V86.876"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <circle
-        cx="4.38608"
-        cy="4.38608"
-        r="4.38608"
-        transform="matrix(-1 0 0 1 68.6758 85.9988)"
-        fill="#343A3F"
-      />
-      <circle cx="46.7572" cy="106.614" r="4.38608" fill="#343A3F" />
-      <line
-        x1="46.9184"
-        y1="56.1741"
-        x2="46.9184"
-        y2="106.614"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <path
-        d="M13.8099 28.1615C15.0025 28.1615 16.1571 28.3208 17.2596 28.614C17.2596 28.5851 17.2596 28.5561 17.2596 28.5271C17.2561 12.772 29.4879 0 44.5765 0C55.7439 0 65.3372 6.99784 69.5774 17.0222C70.777 16.7796 72.0148 16.6493 73.2837 16.6493C83.9657 16.6457 92.6299 25.6925 92.6299 36.85C92.6299 48.0074 83.9657 57.0543 73.2803 57.0543C72.8087 57.0543 72.3441 57.0289 71.883 56.9964H46.2442C45.6894 57.0326 45.1347 57.0543 44.573 57.0543C44.0114 57.0543 43.4567 57.0326 42.9019 56.9964H13.8064C6.18235 56.9964 0.000591536 50.5415 0.000591536 42.5807C0.000591536 34.6199 6.18235 28.1651 13.8064 28.1651L13.8099 28.1615Z"
-        fill="url(#paint0_linear_549_22301)"
-      />
-      <path
-        fillRule="evenodd"
-        clipRule="evenodd"
-        d="M46.9978 17C42.8537 17 39.4923 20.3614 39.4923 24.5054V27.7065H36.3026C35.5832 27.7065 35 28.2897 35 29.0091V43.7902C35 44.5096 35.5832 45.0927 36.3026 45.0927H57.6974C58.4168 45.0927 59 44.5096 59 43.7902V29.0091C59 28.2897 58.4168 27.7065 57.6974 27.7065H54.5032V24.5021C54.5032 20.358 51.1418 17 46.9978 17ZM42.4256 24.5021V27.7065H51.5699V24.5021C51.5699 21.9802 49.5197 19.9299 46.9978 19.9299C44.4759 19.9299 42.4256 21.9802 42.4256 24.5021ZM42.6877 36.1663C43.4316 34.9085 43.4541 34.5208 43.4541 34.5208L43.4496 34.5072C44.1124 35.17 44.2972 36.3421 44.2972 36.3421C44.4054 36.1933 44.6083 35.9048 44.6759 35.7966C45.0411 35.2196 45.3296 34.7372 45.5415 34.088C45.565 34.0159 45.5875 33.9488 45.6089 33.8849C45.7149 33.569 45.7948 33.3306 45.8435 32.9519C45.9022 32.4965 45.8976 31.8564 45.8571 31.3965C45.8526 31.3424 45.8165 31.0674 45.7985 30.9998C46.5694 31.3289 47.3042 31.7797 47.8362 32.4379C48.4493 33.1908 48.8235 34.0925 48.8371 35.0618C48.8381 35.0737 48.84 35.0909 48.8422 35.1113C48.8492 35.1754 48.8596 35.2707 48.8596 35.3323C49.1436 35.1068 49.3059 34.8138 49.4276 34.4712C49.5899 34.0068 49.635 33.4974 49.5899 33.0105C49.5764 32.8482 49.4953 32.4785 49.4637 32.4515C51.515 34.2052 52.1146 36.9823 52.1146 36.9823C52.3625 38.3122 52.1552 39.6242 51.4293 40.7648C49.5539 43.7177 44.4911 43.7718 42.6246 40.7648C41.741 39.3356 41.8221 37.627 42.6877 36.1663Z"
-        fill="white"
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_549_22301"
-          x1="45.999"
-          y1="-1.46507e-06"
-          x2="57.8749"
-          y2="54.4078"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#FFAA7F" />
-          <stop offset="0.489583" stopColor="#FF844F" />
-          <stop offset="1" stopColor="#FF4205" />
-        </linearGradient>
-      </defs>
-    </svg>
-  );
-};
-
-export default Enclave;
