--- src/components/LandingFeatures/features/Interoperability.tsx
@@ -1,78 +1,0 @@
-import styles from '../landing-features.module.scss';
-
-const Interoperability = () => {
-  return (
-    <svg
-      className={styles.feature__logo}
-      width="114"
-      height="112"
-      viewBox="0 0 114 112"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-    >
-      <path
-        d="M11.9872 111.138V99.3621L22.4744 93.1822V105.145L11.9872 111.138ZM11.4872 86.7205L21.9543 92.3279L11.4872 98.496L1.02008 92.3279L11.4872 86.7205ZM10.9872 99.3621V111.138L0.5 105.145V93.1822L10.9872 99.3621Z"
-        fill="#343A3F"
-        stroke="white"
-      />
-      <path
-        d="M27.9872 24.9846V13.2088L38.4744 7.02885V18.9919L27.9872 24.9846ZM27.4872 0.567228L37.9543 6.17461L27.4872 12.3427L17.0201 6.1746L27.4872 0.567228ZM26.9872 13.2088V24.9846L16.5 18.9919V7.02885L26.9872 13.2088Z"
-        fill="#343A3F"
-        stroke="white"
-      />
-      <path
-        d="M102.987 74.9846V63.2088L113.474 57.0288V68.9919L102.987 74.9846ZM102.487 50.5672L112.954 56.1746L102.487 62.3427L92.0201 56.1746L102.487 50.5672ZM101.987 63.2088V74.9846L91.5 68.9919V57.0288L101.987 63.2088Z"
-        fill="#343A3F"
-        stroke="white"
-      />
-      <line
-        x1="32.8164"
-        y1="22.2975"
-        x2="43.5856"
-        y2="39.5282"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <line
-        x1="41.7203"
-        y1="71.1524"
-        x2="22.3279"
-        y2="92.2755"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <line
-        x1="70.03"
-        y1="60.7853"
-        x2="91.5684"
-        y2="61.8622"
-        stroke="#343A3F"
-        strokeWidth="1.2"
-      />
-      <path
-        d="M50.5485 35.1216C51.2566 34.7128 52.1289 34.7128 52.837 35.1216L69.2014 44.5695C69.9094 44.9783 70.3456 45.7338 70.3456 46.5514V65.4474C70.3456 66.265 69.9094 67.0205 69.2014 67.4293L52.837 76.8772C52.1289 77.286 51.2566 77.286 50.5485 76.8772L34.1841 67.4293C33.4761 67.0205 33.0399 66.265 33.0399 65.4474V46.5514C33.0399 45.7338 33.4761 44.9783 34.1841 44.5695L50.5485 35.1216Z"
-        fill="url(#paint0_linear_549_22274)"
-      />
-      <path
-        d="M44.6377 50.1046C44.6377 50.1046 44.5988 50.8434 43.167 53.2534C41.5087 56.0462 41.3529 59.3152 43.0466 62.0443C46.6173 67.7977 56.308 67.689 59.9007 62.0479C61.29 59.8658 61.6882 57.3621 61.2139 54.8186C61.2139 54.8186 60.0644 49.508 56.1389 46.1594C56.2009 46.2115 56.3531 46.9221 56.3823 47.2252C56.4717 48.1567 56.3841 49.1333 56.0717 50.0197C55.8425 50.6711 55.5248 51.2349 54.9877 51.6697C54.9877 51.5177 54.9496 51.2455 54.9487 51.1545C54.9231 49.2994 54.2045 47.5814 53.0338 46.1391C52.0108 44.8806 50.6082 44.0215 49.1313 43.3896C49.1685 43.5169 49.2375 44.0445 49.2472 44.1488C49.3233 45.0246 49.3375 46.2513 49.2225 47.1262C49.1074 48.0012 48.9101 48.4775 48.6419 49.2977C48.2358 50.5411 47.6827 51.4629 46.9818 52.5685C46.8527 52.7718 46.4668 53.3294 46.2606 53.6114C46.2606 53.6114 45.9058 51.3701 44.6386 50.1046H44.6377Z"
-        fill="white"
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_549_22274"
-          x1="39"
-          y1="31.5"
-          x2="47.1551"
-          y2="73.39"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#FFAA7F" />
-          <stop offset="0.489583" stopColor="#FF844F" />
-          <stop offset="1" stopColor="#FF4205" />
-        </linearGradient>
-      </defs>
-    </svg>
-  );
-};
-
-export default Interoperability;
