--- src/components/LandingFeatures/features/Threshold.tsx
@@ -1,80 +1,0 @@
-import styles from '../landing-features.module.scss';
-
-const Threshold = () => {
-  return (
-    <svg
-      width="108"
-      height="112"
-      viewBox="0 0 108 112"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      className={styles.feature__logo}
-    >
-      <path
-        d="M26.0649 55.8534L31.0349 60.3557L33.8627 60.0845L37.1304 57.6548L40.0779 60.0193L43.2316 57.7591L46.2931 59.9541L49.6344 59.9998C51.4189 64.6973 56.3551 67.5925 61.4711 66.64C67.2207 65.5624 71.0142 60.0314 69.9421 54.2822C68.8675 48.5301 63.3391 44.7397 57.59 45.8117C53.9915 46.4809 51.16 48.8994 49.8071 52.027L32.0206 51.687L26.0675 55.8564L26.0649 55.8534ZM60.8632 55.9753C60.6208 54.6803 61.4749 53.436 62.7699 53.1935C64.0649 52.9511 65.3092 53.8052 65.5516 55.1002C65.794 56.3952 64.94 57.6395 63.645 57.8819C62.3499 58.1244 61.1057 57.2703 60.8632 55.9753V55.9753Z"
-        stroke="#343a3f"
-        fill="#343a3f"
-      />
-      <path
-        d="M15.7333 84.7804L25.5284 76.7175C25.5284 76.7175 29.1153 73.8493 35.9019 74.9156L55.726 78.0303C55.726 78.0303 62.1669 77.842 61.2143 83.9046C61.2143 83.9046 60.2152 88.5365 54.6294 87.6589L41.4565 85.5892"
-        stroke="#343a3f"
-        strokeWidth={1.2}
-        strokeMiterlimit="10"
-      />
-      <path
-        d="M27.8493 101.009L32.4294 97.9228C32.4294 97.9228 33.6835 96.8894 39.0044 97.7254L56.6978 100.505C56.6978 100.505 62.6598 100.755 66.7798 96.221L79.6189 81.6823C79.6189 81.6823 83.3089 77.1595 79.7516 73.7357C76.1943 70.312 67.5458 77.795 67.5458 77.795L61.4024 82.6922"
-        stroke="#343a3f"
-        strokeWidth={1.2}
-        strokeMiterlimit="10"
-      />
-      <path
-        d="M29.7526 103.883L12.4971 80.0441L3.28195 86.642L20.4664 110.741L29.7526 103.883Z"
-        fill="url(#paint0_linear_1301_1644)"
-      />
-      <path
-        d="M91.6319 27.2186L81.8368 35.2815C81.8368 35.2815 78.2499 38.1498 71.4633 37.0835L51.6392 33.9687C51.6392 33.9687 45.1983 34.157 46.1509 28.0944C46.1509 28.0944 47.15 23.4625 52.7359 24.3402L65.9088 26.4099"
-        stroke="#343a3f"
-        strokeWidth={1.2}
-        strokeMiterlimit="10"
-      />
-      <path
-        d="M79.5199 10.9911L74.9397 14.0772C74.9397 14.0772 73.6857 15.1106 68.3647 14.2746L50.6714 11.4946C50.6714 11.4946 44.7094 11.2455 40.5894 15.779L27.7503 30.3177C27.7503 30.3177 24.0602 34.8405 27.6175 38.2643C31.1748 41.688 39.8234 34.205 39.8234 34.205L45.9667 29.3078"
-        stroke="#343a3f"
-        strokeWidth={1.2}
-        strokeMiterlimit="10"
-      />
-      <path
-        d="M77.6126 8.11709L94.8682 31.9559L104.083 25.358L86.8988 1.25892L77.6126 8.11709Z"
-        fill="url(#paint1_linear_1301_1644)"
-      />
-      <defs>
-        <linearGradient
-          id="paint0_linear_1301_1644"
-          x1="4.51562"
-          y1="78.79"
-          x2="3.82657"
-          y2="121.292"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#FFAA7F" />
-          <stop offset="0.489583" stopColor="#FF4205" />
-          <stop offset="1" stopColor="#FF4205" />
-        </linearGradient>
-        <linearGradient
-          id="paint1_linear_1301_1644"
-          x1="86.5"
-          y1="2.5"
-          x2="102.072"
-          y2="43.2371"
-          gradientUnits="userSpaceOnUse"
-        >
-          <stop stopColor="#FFAA7F" />
-          <stop offset="0.489583" stopColor="#FF4205" />
-          <stop offset="1" stopColor="#FF4205" />
-        </linearGradient>
-      </defs>
-    </svg>
-  );
-};
-
-export default Threshold;
