--- src/components/LandingFeatures/landing-features-2.module.scss
@@ -1,48 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.tab {
-  &[data-active] {
-    z-index: 1;
-    border-top: 4px solid #000;
-  }
-}
-
-.tabsWrapper {
-  width: 100%;
-  overflow-x: auto;
-  overflow-y: hidden;
-  -webkit-overflow-scrolling: touch;
-  margin: 0;
-  padding: 1rem 0;
-  scrollbar-width: none; /* Firefox */
-
-  /* Add these to ensure the scrolling works */
-  display: flex;
-}
-
-.tabsList {
-  display: inline-flex;
-  flex-wrap: nowrap;
-  gap: 12px;
-  width: max-content;
-  min-width: 100%;
-  padding: 0;
-  margin: 0;
-  min-width: 120%;
-}
-
-.tabsPanel {
-    height: 100%;
-}
-
-.features {
-  &__title {
-    width: 90%;
-    font-size: 1.5rem;
-    padding: 0 .5rem;
-
-    // @include sm {
-    //   font-size: 1.75rem;
-    // }
-  }
-}
