--- src/components/LandingFeatures/landing-features.module.scss
@@ -1,208 +1,0 @@
-@import '../../styles/mixins.scss';
-
-.section {
-  width: 100%;
-  background-color: var(--lit-white);
-  color: var(--lit-gray-700);
-}
-
-.wrapper {
-  max-width: 72rem;
-  margin: 0 auto;
-  padding: 5rem 1rem;
-
-  @include sm {
-    padding-left: 1.5rem;
-    padding-right: 1.5rem;
-  }
-
-  @include lg {
-    padding: 8rem 4rem;
-  }
-}
-
-.header {
-  max-width: 56rem;
-  text-align: center;
-  margin: 0 auto;
-}
-
-.badge {
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 1rem;
-
-  @include lg {
-    margin-bottom: 1.5rem;
-  }
-
-  &__copy {
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    color: var(--lit-pink);
-    background-color: var(--lit-pink);
-    background-image: linear-gradient(
-      to bottom,
-      var(--lit-orange),
-      var(--lit-pink)
-    );
-    background-size: 100%;
-    -webkit-background-clip: text;
-    -moz-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    -moz-text-fill-color: transparent;
-
-    @include lg {
-      font-size: 1rem;
-      line-height: 1.5rem;
-    }
-  }
-}
-
-.title {
-  font-size: 1.875rem;
-  line-height: 2.25rem;
-  font-weight: bold;
-  color: var(--lit-gray-900);
-  background-color: var(--lit-gray-900);
-  background-image: linear-gradient(
-    to bottom,
-    var(--lit-gray-900),
-    var(--lit-gray-700)
-  );
-  background-size: 100%;
-  -webkit-background-clip: text;
-  -moz-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  -moz-text-fill-color: transparent;
-
-  @include xs {
-    font-size: 2.25rem;
-    line-height: 2.5rem;
-  }
-
-  @include lg {
-    font-size: 3rem;
-    line-height: 1;
-  }
-
-  &__break {
-    display: inline;
-
-    @include md {
-      display: block;
-    }
-  }
-}
-
-.subtitle {
-  font-size: 1rem;
-  line-height: 1.5rem;
-  font-weight: 300;
-  margin-top: 0.75rem;
-  margin-bottom: 1rem;
-
-  @include lg {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    margin-top: 1rem;
-  }
-}
-.grid {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 2rem;
-  align-items: stretch;
-  margin-top: 2rem;
-  margin-bottom: 1rem;
-
-  @include sm {
-    grid-template-columns: repeat(2, 1fr);
-    margin-top: 3rem;
-  }
-
-  @include lg {
-    grid-template-columns: repeat(3, 1fr);
-  }
-}
-
-.feature {
-  &__logo {
-    width: auto;
-    max-height: 5rem;
-    margin-bottom: 1.5rem;
-
-    @include lg {
-      max-height: 6rem;
-    }
-  }
-
-  &__title {
-    font-size: 1.25rem;
-    line-height: 1.75rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-    color: var(--lit-gray-900);
-    background-color: var(--lit-gray-900);
-    background-image: linear-gradient(
-      to bottom,
-      var(--lit-gray-900),
-      var(--lit-gray-700)
-    );
-    background-size: 100%;
-    -webkit-background-clip: text;
-    -moz-background-clip: text;
-    -webkit-text-fill-color: transparent;
-    -moz-text-fill-color: transparent;
-
-    @include lg {
-      font-size: 1.5rem;
-      line-height: 2rem;
-    }
-  }
-
-  &__copy {
-    letter-spacing: 0.025em;
-    font-size: 0.875rem;
-    line-height: 1.25rem;
-    font-weight: 300;
-
-    @include lg {
-      font-size: 1rem;
-      line-height: 1.5rem;
-    }
-  }
-
-  &__highlight {
-    color: var(--lit-gray-900);
-    font-weight: 400;
-  }
-}
-
-.feature:last-of-type {
-  border-right: none;
-}
-
-.cta-btn {
-  display: inline-flex;
-  justify-content: center;
-  background-color: var(--lit-gray-900);
-  color: var(--lit-gray-50);
-  border: 2px solid var(--lit-gray-900);
-  border-radius: 999px;
-  font-size: 0.875rem;
-  line-height: 1.25rem;
-  padding: 0.625rem 1rem 0.5rem;
-  transition: background-color 0.2s ease, border-color 0.2s ease;
-  margin-right: 0.5rem;
-
-  &:hover {
-    background-color: var(--lit-orange);
-    border-color: var(--lit-orange);
-  }
-
-  @include sm {
-    font-size: 1rem;
-    line-height: 1.5rem;
-  }
-}
